arm: dts: aspeed: Disable UCD90160 in the Zaius dts
Author: John Wang

Merged from Witherspoon.

This is a work-around for the lack of PGOOD issue:

Binding the UCD9000 driver to the device (and possibly the sequence of
unbinding, poking the UCD90160 via i2cset, then re-binding the UCD9000
driver) appears to have some negative effects.  Instrumenting the
transactions with reads of the PMBus status tree lead to the following
observations under "normal" operation.

Reduce our problems by communicating with the UCD90160 as little as
possible.  With the associated userspace change, the only communication
will be applying the VCS work-around.

Change-Id: Ied0e3ec2a8d6f41162476a197df474d041e0f509
Signed-off-by: John Wang <hsienchiang@gmail.com>
Signed-off-by: Doyle Huang <doyle.sy.huang@mail.foxconn.com>